Fix green LEDs in hello world demo
Since the display is now connected to the bottom 4 bits of the GPIO, I
needed to change the GPIO header to increase the mask to 8 bits.
Afterwards the hello world demo is changed to shift through the top 4
bits of the GPIO. Button 0 is still used as a way to pause the shifting.