JTAG TAP correct per implementation
This commit fixes a deficiency of our version of FuseSoC. Previously it
would always use the JTAG TAP, which only works with Sonata and
simulation. This is because flags were not implemented yet. Instead,
this commit uses the synth flag to set whether to use the BSCANE TAP or
the JTAG TAP.

Co-authored-by: Marno van der Maas <mvdmaas+git@lowrisc.org>